added bottom tab navigator skeleton with tabs for Feed and Profile. Successfully navigates between the screens.
