+++ flows/tabs.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { createMaterialBottomTabNavigator } from '@react-navigation/material-bottom-tabs';
+import FeedScreen from './community/feed';
+import ProfileScreen from './account/profile';
+
+const Tab = createMaterialBottomTabNavigator();
+
+export default function Tabs() {
+  return (
+    <Tab.Navigator>
+      <Tab.Screen name="Feed" component={FeedScreen} />
+      <Tab.Screen name="Profile" component={ProfileScreen} />
+    </Tab.Navigator>
+  );
+}
