+++ package.json
@@ -11,13 +11,19 @@
   },
   "dependencies": {
     "@expo/webpack-config": "^18.0.1",
+    "@react-navigation/material-bottom-tabs": "^6.2.16",
+    "@react-navigation/native": "^6.1.7",
+    "@types/react": "~18.0.27",
     "expo": "~48.0.18",
     "expo-status-bar": "~1.4.4",
     "react": "18.2.0",
     "react-native": "0.71.8",
+    "react-native-paper": "^5.9.1",
+    "react-native-safe-area-context": "^4.7.1",
+    "react-native-screens": "^3.23.0",
+    "react-native-vector-icons": "^10.0.0",
     "react-native-web": "~0.18.10",
-    "typescript": "^4.9.4",
-    "@types/react": "~18.0.27"
+    "typescript": "^4.9.5"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
